feat: refactor to functional component and ts
